+++ src/components/CommunityBoard/CreatePostButton.tsx
@@ -1,7 +1,7 @@
 import React, { MouseEvent } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { ROUTER_PATH } from '@/router';
+import { ROUTER_PATH } from '@/routerPath';
 import { Button } from '@chakra-ui/react';
 
 const CreatePostButton: React.FC = () => {
